refactor: CookieExtractor의 패키지 변경해 하위 패키지 참조하도록 수정
